+++ userspace/alibc/src/fchmod.c
@@ -18,6 +18,6 @@ int fchmod(int f, mode_t mode)
       errno = ENOENT;
       return -1;
    }
-   return asyscall(SYS_MODNOD, fds[f]->file, st.st_uid, st.st_gid, mode, 0, 0);
+   return (int)asyscall(SYS_MODNOD, fds[f]->file, st.st_uid, st.st_gid, mode, 0, 0);
 }
